backend/requirements.txt
fastapi
uvicorn
pymysql
psycopg2-binary
pydantic
python-dotenv
pydantic-settings